postRouter.js created and imported into index.js
